+++ packages/atomic/src/components.d.ts
@@ -37,7 +37,6 @@ import { AtomicRecsStore } from "./components/recommendations/atomic-recs-interf
 import { Bindings as Bindings1 } from "./components/search/atomic-search-interface/atomic-search-interface";
 import { AriaLabelGenerator as AriaLabelGenerator1 } from "./components/search/search-box-suggestions/atomic-search-box-instant-results/atomic-search-box-instant-results";
 import { InitializationOptions } from "./components/search/atomic-search-interface/atomic-search-interface";
-import { FollowUpQuestionCandidate, SelectFollowUpQuestionCandidatePayload } from "./components/common/follow-up-questions/follow-up-question-list-common";
 export { AutomaticFacet, CategoryFacetSortCriterion, DateFilterRange, DateRangeRequest, FacetResultsMustMatch, FacetSortCriterion, FoldedResult, GeneratedAnswer, GeneratedAnswerCitation, InlineLink, InteractiveCitation, InteractiveResult, LogLevel as LogLevel1, RangeFacetRangeAlgorithm, RangeFacetSortCriterion, Result, ResultTemplate, ResultTemplateCondition, SearchEngine, SearchStatus } from "@coveo/headless";
 export { CategoryFacet, CommerceEngine, DateFacet, InteractiveProduct, LogLevel, NumericFacet, Product, ProductListing, ProductListingSummaryState, ProductTemplate, ProductTemplateCondition, RegularFacet, Search, SearchSummaryState, Summary } from "@coveo/headless/commerce";
 export { CommerceBindings as Bindings } from "./components/commerce/atomic-commerce-interface/atomic-commerce-interface";
@@ -70,7 +69,6 @@ export { AtomicRecsStore } from "./components/recommendations/atomic-recs-interf
 export { Bindings as Bindings1 } from "./components/search/atomic-search-interface/atomic-search-interface";
 export { AriaLabelGenerator as AriaLabelGenerator1 } from "./components/search/search-box-suggestions/atomic-search-box-instant-results/atomic-search-box-instant-results";
 export { InitializationOptions } from "./components/search/atomic-search-interface/atomic-search-interface";
-export { FollowUpQuestionCandidate, SelectFollowUpQuestionCandidatePayload } from "./components/common/follow-up-questions/follow-up-question-list-common";
 export namespace Components {
     /**
      * The `atomic-aria-live` component notifies screen readers of changes in the search interface.
@@ -1027,13 +1025,6 @@ export namespace Components {
           * The tabs on which the folded result list can be displayed. This property should not be used at the same time as `tabs-excluded`.  Set this property as a stringified JSON array, e.g., ```html  <atomic-folded-result-list tabs-included='["tabIDA", "tabIDB"]'></atomic-folded-result-list snippet> ``` If you don't set this property, the folded result list can be displayed on any tab. Otherwise, the folded result list can only be displayed on the specified tabs.
          */
         "tabsIncluded": string[] | string;
-    }
-    interface AtomicFollowUpQuestion {
-        "i18n": i18n;
-        "question": string;
-    }
-    interface AtomicFollowUpQuestionList {
-        "answerConfigurationId": string;
     }
     /**
      * The `atomic-format-currency` component is used for formatting currencies.
@@ -3678,6 +3669,20 @@ export namespace Components {
          */
         "name": string;
     }
+    interface AtomicTabButton {
+        /**
+          * Whether the tab button is active.
+         */
+        "active": boolean;
+        /**
+          * The label to display on the tab button.
+         */
+        "label": string;
+        /**
+          * Click handler for the tab button.
+         */
+        "select": () => void;
+    }
     /**
      * @alpha 
      */
@@ -3686,6 +3691,8 @@ export namespace Components {
           * Whether to clear the filters when the active tab changes.
          */
         "clearFiltersOnTabChange"?: boolean;
+    }
+    interface AtomicTabManagerBar {
     }
     /**
      * The `atomic-table-element` element defines a table column in a result list.
@@ -3793,29 +3800,10 @@ export namespace Components {
          */
         "withDatePicker": boolean;
     }
-    interface FollowUpQuestionListCommon {
-        "candidates": FollowUpQuestionCandidate[];
-        "i18n": i18n;
-    }
     interface TabBar {
     }
-    interface TabButton {
-        /**
-          * Whether the tab button is active.
-         */
-        "active": boolean;
-        /**
-          * The label to display on the tab button.
-         */
-        "label": string;
-        /**
-          * Click handler for the tab button.
-         */
-        "select": () => void;
-    }
-    interface TabManagerBar {
-    }
     interface TabPopover {
+        "closePopoverOnFocusOut": (event: FocusEvent) => Promise<void>;
         "setButtonVisibility": (isVisible: boolean) => Promise<void>;
         "togglePopover": () => Promise<void>;
     }
@@ -3843,10 +3831,6 @@ export interface AtomicFacetDateInputCustomEvent<T> extends CustomEvent<T> {
 export interface AtomicFacetNumberInputCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLAtomicFacetNumberInputElement;
-}
-export interface AtomicFollowUpQuestionCustomEvent<T> extends CustomEvent<T> {
-    detail: T;
-    target: HTMLAtomicFollowUpQuestionElement;
 }
 export interface AtomicGeneratedAnswerFeedbackModalCustomEvent<T> extends CustomEvent<T> {
     detail: T;
@@ -3923,10 +3907,6 @@ export interface AtomicSmartSnippetFeedbackModalCustomEvent<T> extends CustomEve
 export interface AtomicSmartSnippetSourceCustomEvent<T> extends CustomEvent<T> {
     detail: T;
     target: HTMLAtomicSmartSnippetSourceElement;
-}
-export interface FollowUpQuestionListCommonCustomEvent<T> extends CustomEvent<T> {
-    detail: T;
-    target: HTMLFollowUpQuestionListCommonElement;
 }
 declare global {
     /**
@@ -4430,29 +4410,6 @@ declare global {
     var HTMLAtomicFoldedResultListElement: {
         prototype: HTMLAtomicFoldedResultListElement;
         new (): HTMLAtomicFoldedResultListElement;
-    };
-    interface HTMLAtomicFollowUpQuestionElementEventMap {
-        "select": void;
-    }
-    interface HTMLAtomicFollowUpQuestionElement extends Components.AtomicFollowUpQuestion, HTMLStencilElement {
-        addEventListener<K extends keyof HTMLAtomicFollowUpQuestionElementEventMap>(type: K, listener: (this: HTMLAtomicFollowUpQuestionElement, ev: AtomicFollowUpQuestionCustomEvent<HTMLAtomicFollowUpQuestionElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
-        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
-        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
-        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
-        removeEventListener<K extends keyof HTMLAtomicFollowUpQuestionElementEventMap>(type: K, listener: (this: HTMLAtomicFollowUpQuestionElement, ev: AtomicFollowUpQuestionCustomEvent<HTMLAtomicFollowUpQuestionElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
-        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
-        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
-        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
-    }
-    var HTMLAtomicFollowUpQuestionElement: {
-        prototype: HTMLAtomicFollowUpQuestionElement;
-        new (): HTMLAtomicFollowUpQuestionElement;
-    };
-    interface HTMLAtomicFollowUpQuestionListElement extends Components.AtomicFollowUpQuestionList, HTMLStencilElement {
-    }
-    var HTMLAtomicFollowUpQuestionListElement: {
-        prototype: HTMLAtomicFollowUpQuestionListElement;
-        new (): HTMLAtomicFollowUpQuestionListElement;
     };
     /**
      * The `atomic-format-currency` component is used for formatting currencies.
@@ -6047,6 +6004,12 @@ declare global {
         prototype: HTMLAtomicTabElement;
         new (): HTMLAtomicTabElement;
     };
+    interface HTMLAtomicTabButtonElement extends Components.AtomicTabButton, HTMLStencilElement {
+    }
+    var HTMLAtomicTabButtonElement: {
+        prototype: HTMLAtomicTabButtonElement;
+        new (): HTMLAtomicTabButtonElement;
+    };
     /**
      * @alpha 
      */
@@ -6055,6 +6018,12 @@ declare global {
     var HTMLAtomicTabManagerElement: {
         prototype: HTMLAtomicTabManagerElement;
         new (): HTMLAtomicTabManagerElement;
+    };
+    interface HTMLAtomicTabManagerBarElement extends Components.AtomicTabManagerBar, HTMLStencilElement {
+    }
+    var HTMLAtomicTabManagerBarElement: {
+        prototype: HTMLAtomicTabManagerBarElement;
+        new (): HTMLAtomicTabManagerBarElement;
     };
     /**
      * The `atomic-table-element` element defines a table column in a result list.
@@ -6094,40 +6063,11 @@ declare global {
         prototype: HTMLAtomicTimeframeFacetElement;
         new (): HTMLAtomicTimeframeFacetElement;
     };
-    interface HTMLFollowUpQuestionListCommonElementEventMap {
-        "selectCandidate": SelectFollowUpQuestionCandidatePayload;
-    }
-    interface HTMLFollowUpQuestionListCommonElement extends Components.FollowUpQuestionListCommon, HTMLStencilElement {
-        addEventListener<K extends keyof HTMLFollowUpQuestionListCommonElementEventMap>(type: K, listener: (this: HTMLFollowUpQuestionListCommonElement, ev: FollowUpQuestionListCommonCustomEvent<HTMLFollowUpQuestionListCommonElementEventMap[K]>) => any, options?: boolean | AddEventListenerOptions): void;
-        addEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
-        addEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | AddEventListenerOptions): void;
-        addEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | AddEventListenerOptions): void;
-        removeEventListener<K extends keyof HTMLFollowUpQuestionListCommonElementEventMap>(type: K, listener: (this: HTMLFollowUpQuestionListCommonElement, ev: FollowUpQuestionListCommonCustomEvent<HTMLFollowUpQuestionListCommonElementEventMap[K]>) => any, options?: boolean | EventListenerOptions): void;
-        removeEventListener<K extends keyof DocumentEventMap>(type: K, listener: (this: Document, ev: DocumentEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
-        removeEventListener<K extends keyof HTMLElementEventMap>(type: K, listener: (this: HTMLElement, ev: HTMLElementEventMap[K]) => any, options?: boolean | EventListenerOptions): void;
-        removeEventListener(type: string, listener: EventListenerOrEventListenerObject, options?: boolean | EventListenerOptions): void;
-    }
-    var HTMLFollowUpQuestionListCommonElement: {
-        prototype: HTMLFollowUpQuestionListCommonElement;
-        new (): HTMLFollowUpQuestionListCommonElement;
-    };
     interface HTMLTabBarElement extends Components.TabBar, HTMLStencilElement {
     }
     var HTMLTabBarElement: {
         prototype: HTMLTabBarElement;
         new (): HTMLTabBarElement;
-    };
-    interface HTMLTabButtonElement extends Components.TabButton, HTMLStencilElement {
-    }
-    var HTMLTabButtonElement: {
-        prototype: HTMLTabButtonElement;
-        new (): HTMLTabButtonElement;
-    };
-    interface HTMLTabManagerBarElement extends Components.TabManagerBar, HTMLStencilElement {
-    }
-    var HTMLTabManagerBarElement: {
-        prototype: HTMLTabManagerBarElement;
-        new (): HTMLTabManagerBarElement;
     };
     interface HTMLTabPopoverElement extends Components.TabPopover, HTMLStencilElement {
     }
@@ -6180,8 +6120,6 @@ declare global {
         "atomic-field-condition": HTMLAtomicFieldConditionElement;
         "atomic-focus-trap": HTMLAtomicFocusTrapElement;
         "atomic-folded-result-list": HTMLAtomicFoldedResultListElement;
-        "atomic-follow-up-question": HTMLAtomicFollowUpQuestionElement;
-        "atomic-follow-up-question-list": HTMLAtomicFollowUpQuestionListElement;
         "atomic-format-currency": HTMLAtomicFormatCurrencyElement;
         "atomic-format-number": HTMLAtomicFormatNumberElement;
         "atomic-format-unit": HTMLAtomicFormatUnitElement;
@@ -6329,15 +6267,14 @@ declare global {
         "atomic-sort-dropdown": HTMLAtomicSortDropdownElement;
         "atomic-sort-expression": HTMLAtomicSortExpressionElement;
         "atomic-tab": HTMLAtomicTabElement;
+        "atomic-tab-button": HTMLAtomicTabButtonElement;
         "atomic-tab-manager": HTMLAtomicTabManagerElement;
+        "atomic-tab-manager-bar": HTMLAtomicTabManagerBarElement;
         "atomic-table-element": HTMLAtomicTableElementElement;
         "atomic-text": HTMLAtomicTextElement;
         "atomic-timeframe": HTMLAtomicTimeframeElement;
         "atomic-timeframe-facet": HTMLAtomicTimeframeFacetElement;
-        "follow-up-question-list-common": HTMLFollowUpQuestionListCommonElement;
         "tab-bar": HTMLTabBarElement;
-        "tab-button": HTMLTabButtonElement;
-        "tab-manager-bar": HTMLTabManagerBarElement;
         "tab-popover": HTMLTabPopoverElement;
     }
 }
@@ -7260,14 +7197,6 @@ declare namespace LocalJSX {
           * The tabs on which the folded result list can be displayed. This property should not be used at the same time as `tabs-excluded`.  Set this property as a stringified JSON array, e.g., ```html  <atomic-folded-result-list tabs-included='["tabIDA", "tabIDB"]'></atomic-folded-result-list snippet> ``` If you don't set this property, the folded result list can be displayed on any tab. Otherwise, the folded result list can only be displayed on the specified tabs.
          */
         "tabsIncluded"?: string[] | string;
-    }
-    interface AtomicFollowUpQuestion {
-        "i18n": i18n;
-        "onSelect"?: (event: AtomicFollowUpQuestionCustomEvent<void>) => void;
-        "question"?: string;
-    }
-    interface AtomicFollowUpQuestionList {
-        "answerConfigurationId": string;
     }
     /**
      * The `atomic-format-currency` component is used for formatting currencies.
@@ -9818,6 +9747,20 @@ declare namespace LocalJSX {
          */
         "name": string;
     }
+    interface AtomicTabButton {
+        /**
+          * Whether the tab button is active.
+         */
+        "active"?: boolean;
+        /**
+          * The label to display on the tab button.
+         */
+        "label": string;
+        /**
+          * Click handler for the tab button.
+         */
+        "select": () => void;
+    }
     /**
      * @alpha 
      */
@@ -9826,6 +9769,8 @@ declare namespace LocalJSX {
           * Whether to clear the filters when the active tab changes.
          */
         "clearFiltersOnTabChange"?: boolean;
+    }
+    interface AtomicTabManagerBar {
     }
     /**
      * The `atomic-table-element` element defines a table column in a result list.
@@ -9933,28 +9878,7 @@ declare namespace LocalJSX {
          */
         "withDatePicker"?: boolean;
     }
-    interface FollowUpQuestionListCommon {
-        "candidates"?: FollowUpQuestionCandidate[];
-        "i18n": i18n;
-        "onSelectCandidate"?: (event: FollowUpQuestionListCommonCustomEvent<SelectFollowUpQuestionCandidatePayload>) => void;
-    }
     interface TabBar {
-    }
-    interface TabButton {
-        /**
-          * Whether the tab button is active.
-         */
-        "active"?: boolean;
-        /**
-          * The label to display on the tab button.
-         */
-        "label": string;
-        /**
-          * Click handler for the tab button.
-         */
-        "select": () => void;
-    }
-    interface TabManagerBar {
     }
     interface TabPopover {
     }
@@ -10003,8 +9927,6 @@ declare namespace LocalJSX {
         "atomic-field-condition": AtomicFieldCondition;
         "atomic-focus-trap": AtomicFocusTrap;
         "atomic-folded-result-list": AtomicFoldedResultList;
-        "atomic-follow-up-question": AtomicFollowUpQuestion;
-        "atomic-follow-up-question-list": AtomicFollowUpQuestionList;
         "atomic-format-currency": AtomicFormatCurrency;
         "atomic-format-number": AtomicFormatNumber;
         "atomic-format-unit": AtomicFormatUnit;
@@ -10152,15 +10074,14 @@ declare namespace LocalJSX {
         "atomic-sort-dropdown": AtomicSortDropdown;
         "atomic-sort-expression": AtomicSortExpression;
         "atomic-tab": AtomicTab;
+        "atomic-tab-button": AtomicTabButton;
         "atomic-tab-manager": AtomicTabManager;
+        "atomic-tab-manager-bar": AtomicTabManagerBar;
         "atomic-table-element": AtomicTableElement;
         "atomic-text": AtomicText;
         "atomic-timeframe": AtomicTimeframe;
         "atomic-timeframe-facet": AtomicTimeframeFacet;
-        "follow-up-question-list-common": FollowUpQuestionListCommon;
         "tab-bar": TabBar;
-        "tab-button": TabButton;
-        "tab-manager-bar": TabManagerBar;
         "tab-popover": TabPopover;
     }
 }
@@ -10384,8 +10305,6 @@ declare module "@stencil/core" {
              * The `atomic-folded-result-list` component is responsible for displaying folded query results, by applying one or more result templates for up to three layers (i.e., to the result, child and grandchild).
              */
             "atomic-folded-result-list": LocalJSX.AtomicFoldedResultList & JSXBase.HTMLAttributes<HTMLAtomicFoldedResultListElement>;
-            "atomic-follow-up-question": LocalJSX.AtomicFollowUpQuestion & JSXBase.HTMLAttributes<HTMLAtomicFollowUpQuestionElement>;
-            "atomic-follow-up-question-list": LocalJSX.AtomicFollowUpQuestionList & JSXBase.HTMLAttributes<HTMLAtomicFollowUpQuestionListElement>;
             /**
              * The `atomic-format-currency` component is used for formatting currencies.
              * The numerical format of compatible parents will be set according to the currency property of this component.
@@ -11024,10 +10943,12 @@ declare module "@stencil/core" {
              */
             "atomic-sort-expression": LocalJSX.AtomicSortExpression & JSXBase.HTMLAttributes<HTMLAtomicSortExpressionElement>;
             "atomic-tab": LocalJSX.AtomicTab & JSXBase.HTMLAttributes<HTMLAtomicTabElement>;
+            "atomic-tab-button": LocalJSX.AtomicTabButton & JSXBase.HTMLAttributes<HTMLAtomicTabButtonElement>;
             /**
              * @alpha 
              */
             "atomic-tab-manager": LocalJSX.AtomicTabManager & JSXBase.HTMLAttributes<HTMLAtomicTabManagerElement>;
+            "atomic-tab-manager-bar": LocalJSX.AtomicTabManagerBar & JSXBase.HTMLAttributes<HTMLAtomicTabManagerBarElement>;
             /**
              * The `atomic-table-element` element defines a table column in a result list.
              */
@@ -11046,10 +10967,7 @@ declare module "@stencil/core" {
              * An `atomic-timeframe-facet` displays a facet of the results for the current query as date intervals.
              */
             "atomic-timeframe-facet": LocalJSX.AtomicTimeframeFacet & JSXBase.HTMLAttributes<HTMLAtomicTimeframeFacetElement>;
-            "follow-up-question-list-common": LocalJSX.FollowUpQuestionListCommon & JSXBase.HTMLAttributes<HTMLFollowUpQuestionListCommonElement>;
             "tab-bar": LocalJSX.TabBar & JSXBase.HTMLAttributes<HTMLTabBarElement>;
-            "tab-button": LocalJSX.TabButton & JSXBase.HTMLAttributes<HTMLTabButtonElement>;
-            "tab-manager-bar": LocalJSX.TabManagerBar & JSXBase.HTMLAttributes<HTMLTabManagerBarElement>;
             "tab-popover": LocalJSX.TabPopover & JSXBase.HTMLAttributes<HTMLTabPopoverElement>;
         }
     }