+++ packages/headless/src/controllers/commerce/context/cart/headless-cart.ssr.ts
@@ -1,4 +1,5 @@
-import {NonRecommendationControllerDefinitionWithProps} from '../../../../app/commerce-ssr-engine/types/common.js';
+import {UniversalControllerDefinitionWithProps} from '../../../../app/commerce-ssr-engine/types/common.js';
+import {Kind} from '../../../../app/commerce-ssr-engine/types/kind.js';
 import {Cart, buildCart, CartInitialState} from './headless-cart.js';
 
 export type {CartState, CartItem, CartProps} from './headless-cart.js';
@@ -9,10 +10,7 @@ export interface CartBuildProps {
 }
 
 export interface CartDefinition
-  extends NonRecommendationControllerDefinitionWithProps<
-    Cart,
-    CartBuildProps
-  > {}
+  extends UniversalControllerDefinitionWithProps<Cart, CartBuildProps> {}
 
 /**
  * Defines a `Cart` controller instance.
@@ -27,7 +25,12 @@ export function defineCart(): CartDefinition {
     listing: true,
     search: true,
     standalone: true,
-    buildWithProps: (engine, props) =>
-      buildCart(engine, {initialState: props.initialState}),
+    recommendation: true,
+    buildWithProps: (engine, props) => {
+      return {
+        ...buildCart(engine, {initialState: props.initialState}),
+        _kind: Kind.Cart,
+      };
+    },
   };
 }