+++ packages/headless/src/controllers/commerce/context/headless-context.ssr.ts
@@ -1,4 +1,5 @@
-import {NonRecommendationControllerDefinitionWithProps} from '../../../app/commerce-ssr-engine/types/common.js';
+import {UniversalControllerDefinitionWithProps} from '../../../app/commerce-ssr-engine/types/common.js';
+import {Kind} from '../../../app/commerce-ssr-engine/types/kind.js';
 import {
   Context,
   buildContext,
@@ -11,10 +12,7 @@ export type {ContextState, Context, ContextProps} from './headless-context.js';
 export type {View, UserLocation, ContextOptions};
 
 export interface ContextDefinition
-  extends NonRecommendationControllerDefinitionWithProps<
-    Context,
-    ContextOptions
-  > {}
+  extends UniversalControllerDefinitionWithProps<Context, ContextOptions> {}
 
 /**
  * Defines a `Context` controller instance.
@@ -29,6 +27,12 @@ export function defineContext(): ContextDefinition {
     listing: true,
     search: true,
     standalone: true,
-    buildWithProps: (engine, props) => buildContext(engine, {options: props}),
+    recommendation: true,
+    buildWithProps: (engine, props) => {
+      return {
+        ...buildContext(engine, {options: props}),
+        _kind: Kind.Context,
+      };
+    },
   };
 }