+++ packages/headless/src/controllers/commerce/recommendations/headless-recommendations.ssr.ts
@@ -1,5 +1,12 @@
-import {RecommendationOnlyControllerDefinitionWithoutProps} from '../../../app/commerce-ssr-engine/types/common.js';
-import {RecommendationsState} from '../recommendations/headless-recommendations.js';
+import {
+  recommendationInternalOptionKey,
+  RecommendationOnlyControllerDefinitionWithProps,
+} from '../../../app/commerce-ssr-engine/types/common.js';
+import {Kind} from '../../../app/commerce-ssr-engine/types/kind.js';
+import {
+  RecommendationsOptions,
+  RecommendationsState,
+} from '../recommendations/headless-recommendations.js';
 import {
   RecommendationsProps,
   Recommendations,
@@ -12,11 +19,15 @@ export type {Recommendations, RecommendationsState};
  * @internal
  * */
 export type RecommendationsDefinitionMeta = {
-  options: {} & RecommendationsProps['options'];
+  [recommendationInternalOptionKey]: {} & RecommendationsProps['options'];
 };
 
 export interface RecommendationsDefinition
-  extends RecommendationOnlyControllerDefinitionWithoutProps<Recommendations> {}
+  extends RecommendationOnlyControllerDefinitionWithProps<
+    Recommendations,
+    Partial<RecommendationsOptions>
+  > {}
+
 /**
  * @internal
  * Defines a `Recommendations` controller instance.
@@ -30,9 +41,27 @@ export function defineRecommendations(
 ): RecommendationsDefinition & RecommendationsDefinitionMeta {
   return {
     recommendation: true,
-    options: {
+    [recommendationInternalOptionKey]: {
       ...props.options,
     },
-    build: (engine) => buildRecommendations(engine, props),
+    buildWithProps: (
+      engine,
+      options: Omit<RecommendationsOptions, 'slotId'>
+    ) => {
+      const staticOptions = props.options;
+      const controller = buildRecommendations(engine, {
+        options: {...staticOptions, ...options},
+      });
+      const copy = Object.defineProperties(
+        {},
+        Object.getOwnPropertyDescriptors(controller)
+      );
+
+      Object.defineProperty(copy, '_kind', {
+        value: Kind.Recommendations,
+      });
+
+      return copy as typeof controller & {_kind: Kind.Recommendations};
+    },
   };
 }