+++ packages/samples/headless-commerce-react/src/components/products-per-page/products-per-page.tsx
@@ -10,20 +10,24 @@ export default function ProductsPerPage(props: IProductsPerPageProps) {
 
   const [state, setState] = useState(controller.state);
 
+  const options = [5, 10, 20, 50, 0];
+
   useEffect(() => {
     controller.subscribe(() => setState(controller.state));
   }, [controller]);
 
-  const options = [5, 10, 20, 50];
   return (
     <span className="ProductsPerPage">
       <span className="ProductsPerPageLabel">Products per page:</span>
       {options.map((pageSize) => {
         const id = `page-size-${pageSize}`;
         return (
-          <span key={pageSize}>
+          <span key={id}>
             <input
-              checked={state.pageSize === pageSize}
+              checked={
+                state.pageSize === pageSize ||
+                (pageSize === 0 && !options.includes(state.pageSize))
+              }
               id={id}
               name={`pageSize-${pageSize}`}
               onChange={() => controller.setPageSize(pageSize)}
@@ -31,26 +35,11 @@ export default function ProductsPerPage(props: IProductsPerPageProps) {
               value={pageSize}
             />
             <label className="ProductsPerPageOption" htmlFor={id}>
-              {pageSize}
+              {pageSize === 0 ? 'Default' : pageSize}
             </label>
           </span>
         );
       })}
-      <span className="ProductsPerPageOptionOther">
-        <input
-          checked={state.pageSize === 0 || !options.includes(state.pageSize)}
-          id="page-size-other"
-          name="page"
-          onChange={() => controller.setPageSize(0)}
-          type="radio"
-          value={state.pageSize}
-        />
-        <label htmlFor="page-size-other" key={0}>
-          {state.pageSize !== 0 && !options.includes(state.pageSize)
-            ? `Other (${state.pageSize})`
-            : 'Default'}
-        </label>
-      </span>
     </span>
   );
 }